+++ runtime/runtime/Input.cpp
@@ -2,29 +2,23 @@
 
 #include "runtime/platform/Platform.h"
 #include "runtime/platform/events/Event.h"
-#include "runtime/shared/Types.h"
 
 #include <bitset>
 
 namespace basalt::input {
 namespace {
 
-constexpr i32 NUM_KEYS = static_cast<i32>(Key::NumberOfKeys);
-constexpr i32 NUM_MOUSE_BUTTONS = static_cast<i32>(
-  MouseButton::NumberOfButtons
-);
-
-std::bitset<NUM_KEYS> sKeyStates;
-std::bitset<NUM_MOUSE_BUTTONS> sButtonStates;
+std::bitset<KEY_COUNT> sKeyStates;
+std::bitset<MOUSE_BUTTON_COUNT> sButtonStates;
 math::Vec2i32 sMousePos;
 
 void on_key_pressed(const platform::KeyPressedEvent& event) {
-  const auto index = static_cast<i32>(event.mKey);
+  const auto index = enum_cast(event.mKey);
   sKeyStates[index] = true;
 }
 
 void on_key_released(const platform::KeyReleasedEvent& event) {
-  const auto index = static_cast<i32>(event.mKey);
+  const auto index = enum_cast(event.mKey);
   sKeyStates[index] = false;
 }
 
@@ -33,12 +27,12 @@ void on_mouse_moved(const platform::MouseMovedEvent& event) {
 }
 
 void on_button_pressed(const platform::MouseButtonPressedEvent& event) {
-  const auto index = static_cast<i32>(event.mButton);
+  const auto index = enum_cast(event.mButton);
   sButtonStates[index] = true;
 }
 
 void on_button_released(const platform::MouseButtonReleasedEvent& event) {
-  const auto index = static_cast<i32>(event.mButton);
+  const auto index = enum_cast(event.mButton);
   sButtonStates[index] = false;
 }
 
@@ -58,16 +52,16 @@ void init() {
   add_event_listener(&platform_event_callback);
 }
 
-auto is_key_pressed(Key key) -> bool {
-  return sKeyStates[static_cast<i32>(key)];
+auto is_key_pressed(const Key key) -> bool {
+  return sKeyStates[enum_cast(key)];
 }
 
 auto get_mouse_pos() -> const math::Vec2i32& {
   return sMousePos;
 }
 
-auto is_mouse_button_pressed(MouseButton button) -> bool {
-  return sButtonStates[static_cast<i32>(button)];
+auto is_mouse_button_pressed(const MouseButton button) -> bool {
+  return sButtonStates[enum_cast(button)];
 }
 
 } // namespace basalt::input